price, selling_price and discount now working in navigation
This is not a permanent solution since it involves inflating
products and adding selling_price and discount_percent to each item and
then pushing each product into an array which is then passed as a token
to the template. This is not a good solution for long-term but does at
least allow us to continue work on the demo shop whilst a cleaner
schema-based solution is sought.